+++ CHANGES.md
@@ -2,4 +2,5 @@
 =======
 
 * automatically load Customer Token from the URL if not available in a cookie
-* coupon campaign methods added.
+* coupon campaign methods added
+* split batch calls into multiple HTTP requests if there are too many of them