+++ CHANGES.md
@@ -4,3 +4,4 @@
 * automatically load Customer Token from the URL if not available in a cookie
 * coupon campaign methods added
 * split batch calls into multiple HTTP requests if there are too many of them
+* change url for JS api endpoint